Rename without upper case for file to avoid colision with class name. This is needed for spikeinterface metapackage.
